move ConfiguredMobEffect class to standalone class
